requirements.txt
Django==2.2.6
djangorestframework==3.10.3
Pillow==6.2.1
psycopg2-binary==2.8.4
pyodbc==4.0.27
python-decouple==3.1
pytz==2019.3
sqlparse==0.3.0
